feat(component): separate semi modifiers
- 결국 모서리에서 바뀔 것들과, 선에서 바뀔 수 있는 것들은 명확히 분리될 수 있어야 한다.
- 이를 통해 UX 상에서 모서리에서의 바뀌는 옵션 값으로 인해 선의 속성들이 정의되는 페인 포인트를 줄일 수 있다.
