+++ apps/web/hooks/index.ts
@@ -2,3 +2,4 @@ export { useToast } from './useToast';
 export { useCreateBlockGroupsStore } from './useCreateBlockGroupsStore';
 export { useBlockGroupsAtom } from './useBlockGroupsAtom';
 export { useBorderMatrix } from './useBorderMatrix';
+export { useBorderModifier } from './useBorderModifier';
