+++ apps/web/templates/layouts/sidebar/right/ActivedBlockBorderModifier.tsx
@@ -1,9 +1,8 @@
-import React, { FormEvent, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { useTheme } from '@emotion/react';
 
 import {
-  Border,
   DefaultBox,
   DefaultDivider,
   DefaultHStack,
@@ -18,8 +17,7 @@ import {
   concurrentlyActivedSections,
 } from '@atoms/blockBorderAtom';
 
-import { useBlockGroupsAtom } from '@hooks/useBlockGroupsAtom';
-import { useBorderMatrix } from '@hooks/useBorderMatrix';
+import { useBorderMatrix, useBorderModifier } from '@hooks/index';
 
 import { TemplatedColorInputWithTitlePresenter } from './TemplatedColorInputWithTitle';
 import { TemplatedInputWithTitlePresenter } from './TemplatedInputWithTitlePresenter';
@@ -126,7 +124,7 @@ export function BorderMatrix() {
   const [activedState, setActivedState] = useState(initialmatrixState);
 
   const activedAll = useMemo(() => {
-    return Object.values(activedState).every((v) => !!v);
+    return Object.values(activedState).every((v) => v);
   }, [activedState]);
 
   const setActiveSections = useCallback(() => {
@@ -254,250 +252,116 @@ export function BorderMatrix() {
   );
 }
 
+/**
+ * @description
+ * INFO: 보더를 설정함에 있어 모서리일 경우랑 선의 경우를 다르게 인터페이스를 가져갈 필요가 있었다.
+ * 이를 좀 더 유연하게 설계하기 위해 팩토리 메서드 패턴을 사용했다.
+ */
+export function BorderSemiModifierFactory() {
+  const { blockBorderState } = useBorderMatrix();
+
+  return blockBorderState.activeBorder in EdgeDirectionsContstants ? (
+    <EdgeModifier />
+  ) : (
+    <LineModifier />
+  );
+}
+
+export function EdgeModifier() {
+  const { activeSectionBorderRadius, setBorderRadiusMiddleWare } = useBorderModifier();
+
+  return (
+    <DefaultVStack spacing={2} paddingTop="4px">
+      <DefaultHStack spacing={2}>
+        <TemplatedInputWithTitlePresenter
+          direction="vertical"
+          inputWidth="42px"
+          title="둥글기"
+          placeholder="입력"
+          value={activeSectionBorderRadius()}
+          onChange={setBorderRadiusMiddleWare}
+        />
+      </DefaultHStack>
+    </DefaultVStack>
+  );
+}
+
+export function LineModifier() {
+  const {
+    activeSectionBorderColor,
+    activeSectionBorderOpacity,
+    activeSectionBorderRadius,
+    activeSectionBorderStyle,
+    activeSectionBorderWidth,
+    setBorderMiddleware,
+    setBorderRadiusMiddleWare,
+  } = useBorderModifier();
+  return (
+    <DefaultVStack spacing={2} paddingTop="4px">
+      <DefaultHStack spacing={2}>
+        <TemplatedInputWithTitlePresenter
+          direction="vertical"
+          inputWidth="42px"
+          title="두께"
+          placeholder="입력"
+          value={activeSectionBorderWidth()}
+          onChange={(e) => setBorderMiddleware(e, 'width')}
+        />
+
+        <TemplatedInputWithTitlePresenter
+          direction="vertical"
+          inputWidth="42px"
+          title="둥글기"
+          placeholder="입력"
+          value={activeSectionBorderRadius()}
+          onChange={setBorderRadiusMiddleWare}
+        />
+
+        <TemplatedInputWithTitlePresenter
+          direction="vertical"
+          inputWidth="48px"
+          title="스타일"
+          placeholder="입력"
+          value={activeSectionBorderStyle()}
+          onChange={(e) => setBorderMiddleware(e, 'style')}
+        />
+      </DefaultHStack>
+
+      <DefaultHStack spacing={2}>
+        <TemplatedColorInputWithTitlePresenter
+          direction="vertical"
+          width="24px"
+          title="색상"
+          value={activeSectionBorderColor()}
+          onChange={(e) => setBorderMiddleware(e, 'color')}
+        />
+
+        <TemplatedInputWithTitlePresenter
+          direction="vertical"
+          inputWidth="60px"
+          title="색상번호"
+          placeholder="입력"
+          value={activeSectionBorderColor()}
+          onChange={(e) => setBorderMiddleware(e, 'color')}
+        />
+
+        <TemplatedInputWithTitlePresenter
+          direction="vertical"
+          inputWidth="48px"
+          title="투명도"
+          placeholder="입력"
+          value={activeSectionBorderOpacity()}
+          onChange={(e) => setBorderMiddleware(e, 'opacity')}
+        />
+      </DefaultHStack>
+    </DefaultVStack>
+  );
+}
+
 /* eslint-disable no-console */
 export function ActivedBlockBorderModifier() {
   const theme = useTheme();
   const { blockBorderState } = useBorderMatrix();
-
-  const {
-    activedBlockGroup,
-    setBlockAllBorderStyle,
-    setBlockAllBorderRadiusStyle,
-    setBlockBorderWidth,
-    setBlockBorderColor,
-    setBlockBorderStyle,
-    setBlockBorderOpacity,
-  } = useBlockGroupsAtom();
-
-  if (!activedBlockGroup?.id) return <div></div>;
-
-  const activeSectionBorderWidth = () => {
-    if (blockBorderState.activeBorder in EdgeDirectionsContstants) return '';
-
-    if (blockBorderState.activeBorder === 'all') {
-      if (
-        Object.values(activedBlockGroup.style.border).every(
-          (v) => v.width === activedBlockGroup.style.border.top.width
-        )
-      ) {
-        return activedBlockGroup.style.border.top.width;
-      } else {
-        return 'mixed';
-      }
-    } else {
-      return activedBlockGroup.style.border[blockBorderState.activeBorder as DirectionsContstants]
-        .width;
-    }
-  };
-
-  const activeSectionBorderColor = () => {
-    if (blockBorderState.activeBorder in EdgeDirectionsContstants) return '';
-
-    if (blockBorderState.activeBorder === 'all') {
-      if (
-        Object.values(activedBlockGroup.style.border).every(
-          (v) => v.color === activedBlockGroup.style.border.top.color
-        )
-      ) {
-        return activedBlockGroup.style.border.top.color;
-      } else {
-        return 'mixed';
-      }
-    } else {
-      return activedBlockGroup.style.border[blockBorderState.activeBorder as DirectionsContstants]
-        .color;
-    }
-  };
-
-  const activeSectionBorderStyle = () => {
-    if (blockBorderState.activeBorder in EdgeDirectionsContstants) return '';
-
-    if (blockBorderState.activeBorder === 'all') {
-      const concurrentlyActivedLines = blockBorderState.concurrentlyActivedSection.filter(
-        (v) => v in DirectionsContstants
-      ) as DirectionsContstants[];
-
-      if (
-        concurrentlyActivedLines.every(
-          (v) =>
-            activedBlockGroup.style.border[v].style === activedBlockGroup.style.border.top.style
-        )
-      ) {
-        return activedBlockGroup.style.border.top.style;
-      } else {
-        return 'mixed';
-      }
-    } else {
-      return activedBlockGroup.style.border[blockBorderState.activeBorder as DirectionsContstants]
-        .style;
-    }
-  };
-
-  const activeSectionBorderOpacity = () => {
-    if (blockBorderState.activeBorder in EdgeDirectionsContstants) return '';
-
-    if (blockBorderState.activeBorder === 'all') {
-      const concurrentlyActivedLines = blockBorderState.concurrentlyActivedSection.filter(
-        (v) => v in DirectionsContstants
-      ) as DirectionsContstants[];
-
-      if (
-        concurrentlyActivedLines.every(
-          (v) =>
-            activedBlockGroup.style.border[v].opacity === activedBlockGroup.style.border.top.opacity
-        )
-      ) {
-        return activedBlockGroup.style.border.top.opacity;
-      } else {
-        return 'mixed';
-      }
-    } else {
-      return activedBlockGroup.style.border[blockBorderState.activeBorder as DirectionsContstants]
-        .opacity;
-    }
-  };
-
-  const activeSectionBorderRadius = () => {
-    const borderRadius = activedBlockGroup.style.borderRadius;
-
-    if (blockBorderState.activeBorder === 'all') {
-      const nowStandardValue = borderRadius.topLeft;
-      if (Object.values(borderRadius).every((v) => v === nowStandardValue)) {
-        return nowStandardValue;
-      } else {
-        return 'mixed';
-      }
-    } else if (blockBorderState.activeBorder in DirectionsContstants) {
-      const concurrentlyActivedEdges = blockBorderState.concurrentlyActivedSection.filter(
-        (v) => v in EdgeDirectionsContstants
-      );
-
-      const nowStandardValue =
-        borderRadius[concurrentlyActivedEdges[0] as EdgeDirectionsContstants];
-
-      if (
-        (concurrentlyActivedEdges as EdgeDirectionsContstants[]).every(
-          (key) => borderRadius[key] === nowStandardValue
-        )
-      ) {
-        return nowStandardValue;
-      } else {
-        return 'mixed';
-      }
-    } else {
-      return borderRadius[blockBorderState.activeBorder as EdgeDirectionsContstants];
-    }
-  };
-
-  const setBorderMiddleware = (e: FormEvent, type: keyof Border) => {
-    const value = (e.target as HTMLInputElement).value;
-
-    if (value === 'mixed') return;
-
-    if (activedBlockGroup.subType === 'text') return;
-
-    if (blockBorderState.activeBorder === 'all') {
-      const nextBorderState = {
-        ...activedBlockGroup.style.border,
-        top: {
-          ...activedBlockGroup.style.border.top,
-          [type]: value,
-        },
-        right: {
-          ...activedBlockGroup.style.border.right,
-          [type]: value,
-        },
-        bottom: {
-          ...activedBlockGroup.style.border.bottom,
-          [type]: value,
-        },
-        left: {
-          ...activedBlockGroup.style.border.left,
-          [type]: value,
-        },
-      };
-
-      setBlockAllBorderStyle({ type: 'block', id: activedBlockGroup.id, border: nextBorderState });
-    } else {
-      if (type === 'width') {
-        if (blockBorderState.activeBorder in DirectionsContstants) {
-          setBlockBorderWidth({
-            subType: activedBlockGroup.subType,
-            type: activedBlockGroup.type,
-            id: activedBlockGroup.id,
-            key: blockBorderState.activeBorder as DirectionsContstants,
-            borderWidth: value,
-          });
-        }
-      }
-
-      if (type === 'color') {
-        setBlockBorderColor({
-          subType: activedBlockGroup.subType,
-          type: activedBlockGroup.type,
-          id: activedBlockGroup.id,
-          key: blockBorderState.activeBorder as DirectionsContstants,
-          borderColor: value,
-        });
-      }
-
-      if (type === 'style') {
-        setBlockBorderStyle({
-          subType: activedBlockGroup.subType,
-          type: activedBlockGroup.type,
-          id: activedBlockGroup.id,
-          key: blockBorderState.activeBorder as DirectionsContstants,
-          borderStyle: value,
-        });
-      }
-
-      if (type === 'opacity') {
-        setBlockBorderOpacity({
-          subType: activedBlockGroup.subType,
-          type: activedBlockGroup.type,
-          id: activedBlockGroup.id,
-          key: blockBorderState.activeBorder as DirectionsContstants,
-          borderOpacity: value,
-        });
-      }
-    }
-  };
-
-  const setBorderRadiusMiddleWare = (e: FormEvent) => {
-    const value = (e.target as HTMLInputElement).value;
-
-    if (value === 'mixed') return;
-
-    if (activedBlockGroup.subType === 'text') return;
-
-    const nextBorderRadius = {
-      ...activedBlockGroup.style.borderRadius,
-    };
-
-    if (blockBorderState.activeBorder === 'all') {
-      nextBorderRadius.topLeft = value;
-      nextBorderRadius.topRight = value;
-      nextBorderRadius.bottomLeft = value;
-      nextBorderRadius.bottomRight = value;
-    } else {
-      if (blockBorderState.activeBorder in DirectionsContstants) {
-        blockBorderState.concurrentlyActivedSection.forEach((section) => {
-          if (section in EdgeDirectionsContstants) {
-            nextBorderRadius[section as EdgeDirectionsContstants] = value;
-          }
-        });
-      } else {
-        nextBorderRadius[blockBorderState.activeBorder as EdgeDirectionsContstants] = value;
-      }
-    }
-
-    setBlockAllBorderRadiusStyle({
-      subType: activedBlockGroup.subType,
-      type: activedBlockGroup.type,
-      id: activedBlockGroup.id,
-      borderRadius: nextBorderRadius,
-    });
-  };
 
   return (
     <>
@@ -509,64 +373,7 @@ export function ActivedBlockBorderModifier() {
         <DefaultHStack spacing={2}>
           <BorderMatrix />
 
-          <DefaultVStack spacing={2} paddingTop="4px">
-            <DefaultHStack spacing={2}>
-              <TemplatedInputWithTitlePresenter
-                direction="vertical"
-                inputWidth="42px"
-                title="두께"
-                placeholder="입력"
-                value={activeSectionBorderWidth()}
-                onChange={(e) => setBorderMiddleware(e, 'width')}
-              />
-
-              <TemplatedInputWithTitlePresenter
-                direction="vertical"
-                inputWidth="42px"
-                title="둥글기"
-                placeholder="입력"
-                value={activeSectionBorderRadius()}
-                onChange={setBorderRadiusMiddleWare}
-              />
-
-              <TemplatedInputWithTitlePresenter
-                direction="vertical"
-                inputWidth="48px"
-                title="스타일"
-                placeholder="입력"
-                value={activeSectionBorderStyle()}
-                onChange={(e) => setBorderMiddleware(e, 'style')}
-              />
-            </DefaultHStack>
-
-            <DefaultHStack spacing={2}>
-              <TemplatedColorInputWithTitlePresenter
-                direction="vertical"
-                width="24px"
-                title="색상"
-                value={activeSectionBorderColor()}
-                onChange={(e) => setBorderMiddleware(e, 'color')}
-              />
-
-              <TemplatedInputWithTitlePresenter
-                direction="vertical"
-                inputWidth="60px"
-                title="색상번호"
-                placeholder="입력"
-                value={activeSectionBorderColor()}
-                onChange={(e) => setBorderMiddleware(e, 'color')}
-              />
-
-              <TemplatedInputWithTitlePresenter
-                direction="vertical"
-                inputWidth="48px"
-                title="투명도"
-                placeholder="입력"
-                value={activeSectionBorderOpacity()}
-                onChange={(e) => setBorderMiddleware(e, 'opacity')}
-              />
-            </DefaultHStack>
-          </DefaultVStack>
+          <BorderSemiModifierFactory />
         </DefaultHStack>
       </DefaultVStack>
